requirements.txt
torch>=1.10.0
#transformers>=4.20.0
pytorch-crf>=0.7.2
numpy>=1.21.0
tqdm>=4.62.0
wikipedia-api>=0.5.4
scikit-learn